# openapi.yaml
openapi: "3.0.3"

info:
  title: API MiniBlog — DevSpark
  description: |
    API REST para gestionar autores y publicaciones del blog de DevSpark.
    Permite crear, leer, actualizar y eliminar autores y publicaciones.
  version: "1.0.0"
  contact:
    name: Jorge Alberto Monsalve

servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
  - url: https://api-miniblog.up.railway.app
    description: Servidor de producción en Railway

tags:
  - name: Autores
    description: Operaciones sobre autores del blog
  - name: Publicaciones
    description: Operaciones sobre publicaciones del blog

# ──────────────────────────────────────────────
# RUTAS
# ──────────────────────────────────────────────
paths:

  # ── AUTORES ──────────────────────────────────

  /api/autores:
    get:
      tags: [Autores]
      summary: Listar todos los autores
      responses:
        "200":
          description: Lista de autores obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Autor"
              example:
                - id: 1
                  nombre: Carlos Pérez
                  email: carlos@devblog.co
                  bio: Desarrollador backend apasionado por Node.js
                  creado_en: "2026-05-12T16:39:13.898Z"
        "500":
          $ref: "#/components/responses/Error500"

    post:
      tags: [Autores]
      summary: Crear un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearAutor"
            example:
              nombre: María López
              email: maria@devblog.co
              bio: Escritora técnica especializada en JavaScript
      responses:
        "201":
          description: Autor creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autor"
        "400":
          $ref: "#/components/responses/Error400"
        "500":
          $ref: "#/components/responses/Error500"

  /api/autores/{id}:
    parameters:
      - $ref: "#/components/parameters/IdRuta"

    get:
      tags: [Autores]
      summary: Obtener un autor por su id
      responses:
        "200":
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autor"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

    put:
      tags: [Autores]
      summary: Actualizar un autor existente
      description: Se pueden enviar uno o más campos. Solo los campos enviados se actualizan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarAutor"
            example:
              bio: Nueva biografía actualizada
      responses:
        "200":
          description: Autor actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Autor"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

    delete:
      tags: [Autores]
      summary: Eliminar un autor
      description: Al eliminar un autor, todas sus publicaciones se eliminan también (CASCADE).
      responses:
        "204":
          description: Autor eliminado exitosamente (sin contenido en la respuesta)
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

  # ── PUBLICACIONES ─────────────────────────────

  /api/publicaciones:
    get:
      tags: [Publicaciones]
      summary: Listar todas las publicaciones
      responses:
        "200":
          description: Lista de publicaciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publicacion"
        "500":
          $ref: "#/components/responses/Error500"

    post:
      tags: [Publicaciones]
      summary: Crear una nueva publicación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearPublicacion"
            example:
              autor_id: 1
              titulo: Mi primer post en el blog
              contenido: Este es el contenido completo del artículo...
              publicado: false
      responses:
        "201":
          description: Publicación creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publicacion"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          description: El autor especificado no existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaError"
        "500":
          $ref: "#/components/responses/Error500"

  /api/publicaciones/autor/{autorId}:
    get:
      tags: [Publicaciones]
      summary: Listar publicaciones de un autor con sus datos
      description: Devuelve todas las publicaciones de un autor incluyendo los datos del autor en cada publicación.
      parameters:
        - name: autorId
          in: path
          required: true
          description: Id del autor
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        "200":
          description: Publicaciones del autor obtenidas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicacionConAutor"
              example:
                - id: 1
                  titulo: Introducción a Node.js
                  contenido: Node.js es un entorno de ejecución...
                  publicado: true
                  creado_en: "2026-05-12T16:39:13.900Z"
                  autor:
                    id: 1
                    nombre: Carlos Pérez
                    email: carlos@devblog.co
                    bio: Desarrollador backend apasionado por Node.js
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

  /api/publicaciones/{id}:
    parameters:
      - $ref: "#/components/parameters/IdRuta"

    get:
      tags: [Publicaciones]
      summary: Obtener una publicación por su id
      responses:
        "200":
          description: Publicación encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publicacion"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

    put:
      tags: [Publicaciones]
      summary: Actualizar una publicación existente
      description: Se pueden enviar uno o más campos. Solo los campos enviados se actualizan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarPublicacion"
            example:
              publicado: true
      responses:
        "200":
          description: Publicación actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publicacion"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

    delete:
      tags: [Publicaciones]
      summary: Eliminar una publicación
      responses:
        "204":
          description: Publicación eliminada exitosamente (sin contenido en la respuesta)
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"

# ──────────────────────────────────────────────
# COMPONENTES REUTILIZABLES
# ──────────────────────────────────────────────
components:

  # ── Parámetros ────────────────────────────────
  parameters:
    IdRuta:
      name: id
      in: path
      required: true
      description: Identificador único del recurso
      schema:
        type: integer
        minimum: 1
      example: 1

  # ── Esquemas de datos ─────────────────────────
  schemas:

    Autor:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único generado automáticamente
          example: 1
        nombre:
          type: string
          description: Nombre completo del autor
          example: Carlos Pérez
        email:
          type: string
          format: email
          description: Email único del autor
          example: carlos@devblog.co
        bio:
          type: string
          nullable: true
          description: Biografía corta del autor
          example: Desarrollador backend apasionado por Node.js
        creado_en:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2026-05-12T16:39:13.898Z"

    CrearAutor:
      type: object
      required: [nombre, email]
      properties:
        nombre:
          type: string
          maxLength: 100
          description: Nombre completo del autor (obligatorio)
          example: María López
        email:
          type: string
          format: email
          maxLength: 150
          description: Email único del autor (obligatorio)
          example: maria@devblog.co
        bio:
          type: string
          maxLength: 1000
          nullable: true
          description: Biografía corta del autor (opcional)
          example: Escritora técnica especializada en JavaScript

    ActualizarAutor:
      type: object
      description: Al menos uno de los campos debe estar presente
      properties:
        nombre:
          type: string
          maxLength: 100
          example: Carlos Pérez Actualizado
        email:
          type: string
          format: email
          maxLength: 150
          example: carlos.nuevo@devblog.co
        bio:
          type: string
          maxLength: 1000
          nullable: true
          example: Nueva biografía actualizada

    Publicacion:
      type: object
      properties:
        id:
          type: integer
          example: 1
        autor_id:
          type: integer
          description: Id del autor que escribió la publicación
          example: 1
        titulo:
          type: string
          example: Introducción a Node.js
        contenido:
          type: string
          example: Node.js es un entorno de ejecución de JavaScript del lado del servidor...
        publicado:
          type: boolean
          description: Indica si la publicación está visible al público
          example: true
        creado_en:
          type: string
          format: date-time
          example: "2026-05-12T16:39:13.900Z"

    PublicacionConAutor:
      type: object
      description: Publicación con los datos completos del autor incluidos
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: Introducción a Node.js
        contenido:
          type: string
          example: Node.js es un entorno de ejecución de JavaScript del lado del servidor...
        publicado:
          type: boolean
          example: true
        creado_en:
          type: string
          format: date-time
          example: "2026-05-12T16:39:13.900Z"
        autor:
          type: object
          properties:
            id:
              type: integer
              example: 1
            nombre:
              type: string
              example: Carlos Pérez
            email:
              type: string
              example: carlos@devblog.co
            bio:
              type: string
              nullable: true
              example: Desarrollador backend apasionado por Node.js

    CrearPublicacion:
      type: object
      required: [autor_id, titulo, contenido]
      properties:
        autor_id:
          type: integer
          minimum: 1
          description: Id del autor (obligatorio, debe existir en la base de datos)
          example: 1
        titulo:
          type: string
          maxLength: 200
          description: Título de la publicación (obligatorio)
          example: Mi primer post en el blog
        contenido:
          type: string
          maxLength: 10000
          description: Contenido completo de la publicación (obligatorio)
          example: Este es el contenido completo del artículo...
        publicado:
          type: boolean
          description: Si está publicado o en borrador (por defecto false)
          example: false

    ActualizarPublicacion:
      type: object
      description: Al menos uno de los campos debe estar presente
      properties:
        autor_id:
          type: integer
          minimum: 1
          example: 2
        titulo:
          type: string
          maxLength: 200
          example: Título actualizado
        contenido:
          type: string
          maxLength: 10000
          example: Contenido actualizado del artículo...
        publicado:
          type: boolean
          example: true

    RespuestaError:
      type: object
      properties:
        error:
          type: string
          description: Descripción del error
          example: Autor no encontrado
        detalles:
          type: array
          description: Lista de errores de validación (solo en errores 400)
          items:
            type: object
            properties:
              campo:
                type: string
                example: nombre
              mensaje:
                type: string
                example: El nombre es obligatorio

  # ── Respuestas reutilizables ──────────────────
  responses:

    Error400:
      description: Datos inválidos en la petición
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespuestaError"
          example:
            error: Datos inválidos
            detalles:
              - campo: nombre
                mensaje: El nombre es obligatorio
              - campo: email
                mensaje: El email es obligatorio

    Error404:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespuestaError"
          example:
            error: Autor no encontrado

    Error500:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespuestaError"
          example:
            error: Error interno del servidor